+++ test/OpenTelemetry.Instrumentation.AWS.Tests/TestAWSClientInstrumentation.cs
@@ -3,6 +3,14 @@
 
 using System.Diagnostics;
 using Amazon;
+using Amazon.Bedrock;
+using Amazon.Bedrock.Model;
+using Amazon.BedrockAgent;
+using Amazon.BedrockAgent.Model;
+using Amazon.BedrockAgentRuntime;
+using Amazon.BedrockAgentRuntime.Model;
+using Amazon.BedrockRuntime;
+using Amazon.BedrockRuntime.Model;
 using Amazon.DynamoDBv2;
 using Amazon.DynamoDBv2.Model;
 using Amazon.Runtime;
@@ -198,6 +206,253 @@ public class TestAWSClientInstrumentation
         }
     }
 
+    [Fact]
+#if NETFRAMEWORK
+    public void TestBedrockGetGuardrailSuccessful()
+#else
+    public async Task TestBedrockGetGuardrailSuccessful()
+#endif
+    {
+        var exportedItems = new List<Activity>();
+
+        var parent = new Activity("parent").Start();
+
+        using (Sdk.CreateTracerProviderBuilder()
+                   .AddXRayTraceId()
+                   .SetSampler(new AlwaysOnSampler())
+                   .AddAWSInstrumentation()
+                   .AddInMemoryExporter(exportedItems)
+                   .Build())
+        {
+            var bedrock = new AmazonBedrockClient(new AnonymousAWSCredentials(), RegionEndpoint.USEast1);
+            string requestId = @"fakerequ-esti-dfak-ereq-uestidfakere";
+            string dummyResponse = "{\"GuardrailId\":\"123456789\"}";
+            CustomResponses.SetResponse(bedrock, dummyResponse, requestId, true);
+            var get_guardrail_req = new GetGuardrailRequest();
+            get_guardrail_req.GuardrailIdentifier = "123456789";
+#if NETFRAMEWORK
+            var response = bedrock.GetGuardrail(get_guardrail_req);
+#else
+            var response = await bedrock.GetGuardrailAsync(get_guardrail_req);
+#endif
+            Assert.Single(exportedItems);
+            Activity awssdk_activity = exportedItems[0];
+
+            this.ValidateAWSActivity(awssdk_activity, parent);
+            this.ValidateBedrockActivityTags(awssdk_activity);
+
+            Assert.Equal(Status.Unset, awssdk_activity.GetStatus());
+            Assert.Equal(requestId, Utils.GetTagValue(awssdk_activity, "aws.requestId"));
+        }
+    }
+
+    [Fact]
+#if NETFRAMEWORK
+    public void TestBedrockRuntimeInvokeModelSuccessful()
+#else
+    public async Task TestBedrockRuntimeInvokeModelSuccessful()
+#endif
+    {
+        var exportedItems = new List<Activity>();
+
+        var parent = new Activity("parent").Start();
+
+        using (Sdk.CreateTracerProviderBuilder()
+                   .AddXRayTraceId()
+                   .SetSampler(new AlwaysOnSampler())
+                   .AddAWSInstrumentation()
+                   .AddInMemoryExporter(exportedItems)
+                   .Build())
+        {
+            var bedrockruntime = new AmazonBedrockRuntimeClient(new AnonymousAWSCredentials(), RegionEndpoint.USEast1);
+            string requestId = @"fakerequ-esti-dfak-ereq-uestidfakere";
+            string dummyResponse = "{}";
+            CustomResponses.SetResponse(bedrockruntime, dummyResponse, requestId, true);
+            var invoke_model_req = new InvokeModelRequest();
+            invoke_model_req.ModelId = "amazon.titan-text-express-v1";
+#if NETFRAMEWORK
+            var response = bedrockruntime.InvokeModel(invoke_model_req);
+#else
+            var response = await bedrockruntime.InvokeModelAsync(invoke_model_req);
+#endif
+            Assert.Single(exportedItems);
+            Activity awssdk_activity = exportedItems[0];
+
+            this.ValidateAWSActivity(awssdk_activity, parent);
+            this.ValidateBedrockRuntimeActivityTags(awssdk_activity);
+
+            Assert.Equal(Status.Unset, awssdk_activity.GetStatus());
+            Assert.Equal(requestId, Utils.GetTagValue(awssdk_activity, "aws.requestId"));
+        }
+    }
+
+    [Fact]
+#if NETFRAMEWORK
+    public void TestBedrockAgentRuntimeInvokeAgentSuccessful()
+#else
+    public async Task TestBedrockAgentRuntimeInvokeAgentSuccessful()
+#endif
+    {
+        var exportedItems = new List<Activity>();
+
+        var parent = new Activity("parent").Start();
+
+        using (Sdk.CreateTracerProviderBuilder()
+                   .AddXRayTraceId()
+                   .SetSampler(new AlwaysOnSampler())
+                   .AddAWSInstrumentation()
+                   .AddInMemoryExporter(exportedItems)
+                   .Build())
+        {
+            var bedrockagentruntime = new AmazonBedrockAgentRuntimeClient(new AnonymousAWSCredentials(), RegionEndpoint.USEast1);
+            string requestId = @"fakerequ-esti-dfak-ereq-uestidfakere";
+            string dummyResponse = "{}";
+            CustomResponses.SetResponse(bedrockagentruntime, dummyResponse, requestId, true);
+            var invoke_agent_req = new InvokeAgentRequest();
+            invoke_agent_req.AgentId = "123456789";
+            invoke_agent_req.AgentAliasId = "testalias";
+            invoke_agent_req.SessionId = "test-session-id";
+            invoke_agent_req.InputText = "sample input text";
+#if NETFRAMEWORK
+            var response = bedrockagentruntime.InvokeAgent(invoke_agent_req);
+#else
+            var response = await bedrockagentruntime.InvokeAgentAsync(invoke_agent_req);
+#endif
+            Assert.Single(exportedItems);
+            Activity awssdk_activity = exportedItems[0];
+
+            this.ValidateAWSActivity(awssdk_activity, parent);
+            this.ValidateBedrockAgentRuntimeActivityTags(awssdk_activity);
+
+            Assert.Equal(Status.Unset, awssdk_activity.GetStatus());
+            Assert.Equal(requestId, Utils.GetTagValue(awssdk_activity, "aws.requestId"));
+        }
+    }
+
+// Agent related operation for Bedrock Agent
+    [Fact]
+#if NETFRAMEWORK
+    public void TestBedrockAgentGetAgentSuccessful()
+#else
+    public async Task TestBedrockAgentGetAgentSuccessful()
+#endif
+    {
+        var exportedItems = new List<Activity>();
+
+        var parent = new Activity("parent").Start();
+
+        using (Sdk.CreateTracerProviderBuilder()
+                   .AddXRayTraceId()
+                   .SetSampler(new AlwaysOnSampler())
+                   .AddAWSInstrumentation()
+                   .AddInMemoryExporter(exportedItems)
+                   .Build())
+        {
+            var bedrockagent = new AmazonBedrockAgentClient(new AnonymousAWSCredentials(), RegionEndpoint.USEast1);
+            string requestId = @"fakerequ-esti-dfak-ereq-uestidfakere";
+            string dummyResponse = "{}";
+            CustomResponses.SetResponse(bedrockagent, dummyResponse, requestId, true);
+            var get_agent_req = new GetAgentRequest();
+            get_agent_req.AgentId = "1234567890";
+#if NETFRAMEWORK
+            var response = bedrockagent.GetAgent(get_agent_req);
+#else
+            var response = await bedrockagent.GetAgentAsync(get_agent_req);
+#endif
+            Assert.Single(exportedItems);
+            Activity awssdk_activity = exportedItems[0];
+
+            this.ValidateAWSActivity(awssdk_activity, parent);
+            this.ValidateBedrockAgentAgentOpsActivityTags(awssdk_activity);
+
+            Assert.Equal(Status.Unset, awssdk_activity.GetStatus());
+            Assert.Equal(requestId, Utils.GetTagValue(awssdk_activity, "aws.requestId"));
+        }
+    }
+
+// Knowledge Base related operation for Bedrock Agent
+    [Fact]
+#if NETFRAMEWORK
+    public void TestBedrockAgentGetKnowledgeBaseSuccessful()
+#else
+    public async Task TestBedrockAgentGetKnowledgeBaseSuccessful()
+#endif
+    {
+        var exportedItems = new List<Activity>();
+
+        var parent = new Activity("parent").Start();
+
+        using (Sdk.CreateTracerProviderBuilder()
+                   .AddXRayTraceId()
+                   .SetSampler(new AlwaysOnSampler())
+                   .AddAWSInstrumentation()
+                   .AddInMemoryExporter(exportedItems)
+                   .Build())
+        {
+            var bedrockagent = new AmazonBedrockAgentClient(new AnonymousAWSCredentials(), RegionEndpoint.USEast1);
+            string requestId = @"fakerequ-esti-dfak-ereq-uestidfakere";
+            string dummyResponse = "{}";
+            CustomResponses.SetResponse(bedrockagent, dummyResponse, requestId, true);
+            var get_knowledge_base_req = new GetKnowledgeBaseRequest();
+            get_knowledge_base_req.KnowledgeBaseId = "1234567890";
+#if NETFRAMEWORK
+            var response = bedrockagent.GetKnowledgeBase(get_knowledge_base_req);
+#else
+            var response = await bedrockagent.GetKnowledgeBaseAsync(get_knowledge_base_req);
+#endif
+            Assert.Single(exportedItems);
+            Activity awssdk_activity = exportedItems[0];
+
+            this.ValidateAWSActivity(awssdk_activity, parent);
+            this.ValidateBedrockAgentKnowledgeBaseOpsActivityTags(awssdk_activity);
+
+            Assert.Equal(Status.Unset, awssdk_activity.GetStatus());
+            Assert.Equal(requestId, Utils.GetTagValue(awssdk_activity, "aws.requestId"));
+        }
+    }
+
+// Data Source related operation for Bedrock Agent
+    [Fact]
+#if NETFRAMEWORK
+    public void TestBedrockAgentGetDataSourceSuccessful()
+#else
+    public async Task TestBedrockAgentGetDataSourceSuccessful()
+#endif
+    {
+        var exportedItems = new List<Activity>();
+
+        var parent = new Activity("parent").Start();
+
+        using (Sdk.CreateTracerProviderBuilder()
+                   .AddXRayTraceId()
+                   .SetSampler(new AlwaysOnSampler())
+                   .AddAWSInstrumentation()
+                   .AddInMemoryExporter(exportedItems)
+                   .Build())
+        {
+            var bedrockagent = new AmazonBedrockAgentClient(new AnonymousAWSCredentials(), RegionEndpoint.USEast1);
+            string requestId = @"fakerequ-esti-dfak-ereq-uestidfakere";
+            string dummyResponse = "{}";
+            CustomResponses.SetResponse(bedrockagent, dummyResponse, requestId, true);
+            var get_data_source_req = new GetDataSourceRequest();
+            get_data_source_req.DataSourceId = "1234567890";
+            get_data_source_req.KnowledgeBaseId = "1234567890";
+#if NETFRAMEWORK
+            var response = bedrockagent.GetDataSource(get_data_source_req);
+#else
+            var response = await bedrockagent.GetDataSourceAsync(get_data_source_req);
+#endif
+            Assert.Single(exportedItems);
+            Activity awssdk_activity = exportedItems[0];
+
+            this.ValidateAWSActivity(awssdk_activity, parent);
+            this.ValidateBedrockAgentDataSourceOpsActivityTags(awssdk_activity);
+
+            Assert.Equal(Status.Unset, awssdk_activity.GetStatus());
+            Assert.Equal(requestId, Utils.GetTagValue(awssdk_activity, "aws.requestId"));
+        }
+    }
+
     private void ValidateAWSActivity(Activity aws_activity, Activity parent)
     {
         Assert.Equal(parent.SpanId, aws_activity.ParentSpanId);
@@ -227,5 +482,78 @@ public class TestAWSClientInstrumentation
         Assert.Equal("aws-api", Utils.GetTagValue(sqs_activity, "rpc.system"));
         Assert.Equal("SQS", Utils.GetTagValue(sqs_activity, "rpc.service"));
         Assert.Equal("SendMessage", Utils.GetTagValue(sqs_activity, "rpc.method"));
+    }
+
+    private void ValidateBedrockActivityTags(Activity bedrock_activity)
+    {
+        Assert.Equal("Bedrock.GetGuardrail", bedrock_activity.DisplayName);
+        Assert.Equal("Bedrock", Utils.GetTagValue(bedrock_activity, "aws.service"));
+        Assert.Equal("GetGuardrail", Utils.GetTagValue(bedrock_activity, "aws.operation"));
+        Assert.Equal("us-east-1", Utils.GetTagValue(bedrock_activity, "aws.region"));
+        Assert.Equal("123456789", Utils.GetTagValue(bedrock_activity, "aws.bedrock.guardrail.id"));
+        Assert.Equal("aws-api", Utils.GetTagValue(bedrock_activity, "rpc.system"));
+        Assert.Equal("Bedrock", Utils.GetTagValue(bedrock_activity, "rpc.service"));
+        Assert.Equal("GetGuardrail", Utils.GetTagValue(bedrock_activity, "rpc.method"));
+    }
+
+    private void ValidateBedrockRuntimeActivityTags(Activity bedrock_activity)
+    {
+        Assert.Equal("Bedrock Runtime.InvokeModel", bedrock_activity.DisplayName);
+        Assert.Equal("Bedrock Runtime", Utils.GetTagValue(bedrock_activity, "aws.service"));
+        Assert.Equal("InvokeModel", Utils.GetTagValue(bedrock_activity, "aws.operation"));
+        Assert.Equal("us-east-1", Utils.GetTagValue(bedrock_activity, "aws.region"));
+        Assert.Equal("amazon.titan-text-express-v1", Utils.GetTagValue(bedrock_activity, "gen_ai.request.model"));
+        Assert.Equal("aws_bedrock", Utils.GetTagValue(bedrock_activity, "gen_ai.system"));
+        Assert.Equal("aws-api", Utils.GetTagValue(bedrock_activity, "rpc.system"));
+        Assert.Equal("Bedrock Runtime", Utils.GetTagValue(bedrock_activity, "rpc.service"));
+        Assert.Equal("InvokeModel", Utils.GetTagValue(bedrock_activity, "rpc.method"));
+    }
+
+    private void ValidateBedrockAgentAgentOpsActivityTags(Activity bedrock_activity)
+    {
+        Assert.Equal("Bedrock Agent.GetAgent", bedrock_activity.DisplayName);
+        Assert.Equal("Bedrock Agent", Utils.GetTagValue(bedrock_activity, "aws.service"));
+        Assert.Equal("GetAgent", Utils.GetTagValue(bedrock_activity, "aws.operation"));
+        Assert.Equal("us-east-1", Utils.GetTagValue(bedrock_activity, "aws.region"));
+        Assert.Equal("1234567890", Utils.GetTagValue(bedrock_activity, "aws.bedrock.agent.id"));
+        Assert.Equal("aws-api", Utils.GetTagValue(bedrock_activity, "rpc.system"));
+        Assert.Equal("Bedrock Agent", Utils.GetTagValue(bedrock_activity, "rpc.service"));
+        Assert.Equal("GetAgent", Utils.GetTagValue(bedrock_activity, "rpc.method"));
+    }
+
+    private void ValidateBedrockAgentKnowledgeBaseOpsActivityTags(Activity bedrock_activity)
+    {
+        Assert.Equal("Bedrock Agent.GetKnowledgeBase", bedrock_activity.DisplayName);
+        Assert.Equal("Bedrock Agent", Utils.GetTagValue(bedrock_activity, "aws.service"));
+        Assert.Equal("GetKnowledgeBase", Utils.GetTagValue(bedrock_activity, "aws.operation"));
+        Assert.Equal("us-east-1", Utils.GetTagValue(bedrock_activity, "aws.region"));
+        Assert.Equal("1234567890", Utils.GetTagValue(bedrock_activity, "aws.bedrock.knowledge_base.id"));
+        Assert.Equal("aws-api", Utils.GetTagValue(bedrock_activity, "rpc.system"));
+        Assert.Equal("Bedrock Agent", Utils.GetTagValue(bedrock_activity, "rpc.service"));
+        Assert.Equal("GetKnowledgeBase", Utils.GetTagValue(bedrock_activity, "rpc.method"));
+    }
+
+    private void ValidateBedrockAgentDataSourceOpsActivityTags(Activity bedrock_activity)
+    {
+        Assert.Equal("Bedrock Agent.GetDataSource", bedrock_activity.DisplayName);
+        Assert.Equal("Bedrock Agent", Utils.GetTagValue(bedrock_activity, "aws.service"));
+        Assert.Equal("GetDataSource", Utils.GetTagValue(bedrock_activity, "aws.operation"));
+        Assert.Equal("us-east-1", Utils.GetTagValue(bedrock_activity, "aws.region"));
+        Assert.Equal("1234567890", Utils.GetTagValue(bedrock_activity, "aws.bedrock.data_source.id"));
+        Assert.Equal("aws-api", Utils.GetTagValue(bedrock_activity, "rpc.system"));
+        Assert.Equal("Bedrock Agent", Utils.GetTagValue(bedrock_activity, "rpc.service"));
+        Assert.Equal("GetDataSource", Utils.GetTagValue(bedrock_activity, "rpc.method"));
+    }
+
+    private void ValidateBedrockAgentRuntimeActivityTags(Activity bedrock_activity)
+    {
+        Assert.Equal("Bedrock Agent Runtime.InvokeAgent", bedrock_activity.DisplayName);
+        Assert.Equal("Bedrock Agent Runtime", Utils.GetTagValue(bedrock_activity, "aws.service"));
+        Assert.Equal("InvokeAgent", Utils.GetTagValue(bedrock_activity, "aws.operation"));
+        Assert.Equal("us-east-1", Utils.GetTagValue(bedrock_activity, "aws.region"));
+        Assert.Equal("123456789", Utils.GetTagValue(bedrock_activity, "aws.bedrock.agent.id"));
+        Assert.Equal("aws-api", Utils.GetTagValue(bedrock_activity, "rpc.system"));
+        Assert.Equal("Bedrock Agent Runtime", Utils.GetTagValue(bedrock_activity, "rpc.service"));
+        Assert.Equal("InvokeAgent", Utils.GetTagValue(bedrock_activity, "rpc.method"));
     }
 }